Item Touching and Interaction.
I changed the TakeableItemMap to a class called ItemMap.  The reasoning
was that you want to be able to have all Items grouped by them being
'Items', and upon interaction, the Entity will know what to do.

One thing that I'm not sure on is the attemptRemoveFrom(ItemTile tile)
being placed in all Items.

I didn't want to put a boolean readyToBeRemoved, because that seems
ambiguous. 

However, for things like Obstacle Item, this is an empty method.  I made
a note of this in the source code.